refactor(direct-control): extract production postconditions
Expand packages/civ7-direct-control/test/production-choice.test.ts to cover both cleared and sticky production blocker classifications, then move the production postcondition helper group into packages/civ7-direct-control/src/play/operations/production-postconditions.ts while keeping the embedded production-choice source owner and wrapper-level request composition in the facade.

Record the new production helper owner boundary in the support modularization OpenSpec tasks and workstream records. This is local package and source relocation proof only, not runtime/live-game proof.

Validation: bun run --cwd packages/civ7-direct-control test -- production-choice.test.ts; bun run --cwd packages/cli test -- game.play.production.test.ts; bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; bun run resources:status; bun run check:cli; bun run test:cli:play; git diff --check.
