+++ openspec/changes/civ7-support-direct-control-modularization/tasks.md
@@ -139,6 +139,8 @@ runtime/direct-control claims.
         facade.
   - [x] 4.6.5 Extract population-placement postcondition helper owner while
         keeping production and wrapper-level request composition in the facade.
+  - [x] 4.6.6 Extract production postcondition helper owner while keeping
+        wrapper-level request composition in the facade.
 - [x] 4.7 Extract settlement/tactical/progression read atoms.
   - [x] 4.7.1 Extract settlement recommendation embedded source owner while
         keeping the public wrapper in the facade.
